Fragments and logic
Refactor Repository
Showing Post
Showing Comments

+++ app/src/main/java/com/simplekjl/tddkt/data/models/Comment.kt
@@ -5,8 +5,8 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Comment(
-    var postId: Float = 0.toFloat(),
-    var id: Float = 0.toFloat(),
+    var postId: Int = 0,
+    var id: Int = 0,
     var name: String? = null,
     var email: String? = null,
     var body: String? = null
